Clean up hashing code
