+++ scattered-collect/Cargo.toml
@@ -25,6 +25,10 @@ name = "scattered-collect-intern-strings"
 path = "examples/intern_strings.rs"
 
 [[example]]
+name = "scattered-collect-map"
+path = "examples/map.rs"
+
+[[example]]
 name = "scattered-collect-slice"
 path = "examples/slice.rs"
